lots
-added reactive plating
-added artificer hull
-made updating items on the knight house also update the loaned items on the knights
-made activeeffects from items work on knights
-made the mechbay tabs remember which tab was selected again

+++ module/actor/actor.js
@@ -491,7 +491,7 @@ export class FortyKActor extends Actor {
                 //if item is equipped and/or not disabled
                 if(proceed){
                     ae.changes.forEach(function(change,i){
-                        let basevalue=parseInt(objectByString(actorData,change.key));
+                        let basevalue=parseFloat(objectByString(actorData,change.key));
 
                         let newvalue=parseFloat(change.value);
                         let path=change.key.split(".");
